MGM: Add a new environment variable called EOS_ARCHIVE_URL which needs to point to the location      where archives will be saved. If this variable is not set then no archive transfers are      possible. Use also the uid and gid of the user creating the archive as elements of the final      path where the archive is saved.

+++ mgm/ProcInterface.cc
@@ -382,8 +382,8 @@ ProcCommand::open (const char* inpath,
   // ---------------------------------------------
   // deal with '&' ... sigh 
   // ---------------------------------------------
-  XrdOucString sinfo=ininfo;
-  for (size_t i=0; i< (size_t)sinfo.length(); i++)
+  XrdOucString sinfo = ininfo;
+  for (int i = 0; i < sinfo.length(); i++)
   {
 
     if (sinfo[i] == '&')
